feat: jwt 토큰 생성 구현 및 github oauth 역할 분리 - third party와 통신하는 계층 생성
