+++ backend/build.gradle
@@ -29,6 +29,9 @@ dependencies {
     annotationProcessor 'org.projectlombok:lombok'
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    testImplementation 'io.rest-assured:rest-assured'
+
+    implementation 'io.jsonwebtoken:jjwt:0.9.1'
 }
 
 test {
