+++ backend/src/main/java/kr/kro/colla/auth/service/AuthService.java
@@ -1,67 +1,24 @@
 package kr.kro.colla.auth.service;
 
-import kr.kro.colla.auth.service.dto.GithubAccessTokenRequest;
-import kr.kro.colla.auth.service.dto.GithubAccessTokenResponse;
+import kr.kro.colla.auth.infrastructure.GithubOAuthManager;
 import kr.kro.colla.auth.service.dto.GithubUserProfileResponse;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.*;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestTemplate;
 
+@RequiredArgsConstructor
 @Service
 public class AuthService {
 
-    @Value("${github.client_id}")
-    private String clientId;
+    private final GithubOAuthManager githubOAuthManager;
+    private final JwtProvider jwtProvider;
 
-    @Value("${github.client_secret}")
-    private String clientSecret;
+    public String githubLogin(String code) {
+        String oAuthAccessToken = this.githubOAuthManager.getOAuthAccessToken(code);
+        GithubUserProfileResponse userProfile = this.githubOAuthManager.getUserProfile(oAuthAccessToken);
 
-    @Value("${github.access_token_url}")
-    private String accessTokenUrl;
+        String jwtAccessToken = this.jwtProvider.createToken(userProfile.getGithubId());
 
-    @Value("${github.user_profile_url}")
-    private String userProfileUrl;
-
-    public GithubUserProfileResponse githubLogin(String code) {
-        String accessToken = getAccessToken(code);
-        GithubUserProfileResponse userProfile = getUserProfile(accessToken);
-
-        return userProfile;
-    }
-
-    public String getAccessToken(String code) {
-        GithubAccessTokenRequest githubAccessTokenRequest = new GithubAccessTokenRequest(clientId, clientSecret, code);
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_JSON);
-
-        HttpEntity<GithubAccessTokenRequest> httpEntity = new HttpEntity<>(githubAccessTokenRequest, headers);
-
-        GithubAccessTokenResponse response = new RestTemplate().exchange(
-                accessTokenUrl,
-                HttpMethod.POST,
-                httpEntity,
-                GithubAccessTokenResponse.class
-        ).getBody();
-
-        return response.getAccessToken();
-    }
-
-    public GithubUserProfileResponse getUserProfile(String accessToken) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Authorization", "Bearer " + accessToken);
-
-        HttpEntity<Void> httpEntity = new HttpEntity<>(headers);
-
-        GithubUserProfileResponse response = new RestTemplate().exchange(
-                userProfileUrl,
-                HttpMethod.GET,
-                httpEntity,
-                GithubUserProfileResponse.class
-        ).getBody();
-
-        return response;
+        return jwtAccessToken;
     }
 
 }
